# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the API documentation for the Goals for Good Web Application. More information can be discovered by viewing the repo on [Github](https://github.com/ahmed-alzahrani/GFG-Webapp). You may also view the repo for the mobile client built using the Flutter SDK [here](https://github.com/ahmed-alzahrani/GFG-Client).",
    "version": "1.0.0",
    "title": "Goals For Good Server"
  },
  "host": "localhost:8080",
  "tags": [
    {
      "name": "Player",
      "description": "Operations concerning player information."
    },
    {
      "name": "Countries",
      "description": "Operations concerning geographic information."
    },
    {
      "name": "Charities",
      "description": "Operations concerning charity information."
    },
    {
      "name": "User",
      "description": "Operations concerning user profiles and subscriptions."
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/player/all": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Retrieves info on all players",
        "operationId": "getAllPlayers",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Player"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/'player/matches/{teamId}'": {
      "get": {
        "tags": [
          "Player"
        ],
        "summary": "Retrieves a player's matches",
        "operationId": "getPlayerMatches",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "description": "id of team the player plays for",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Match"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/countries": {
      "get": {
        "tags": [
          "Countries"
        ],
        "summary": "Retrieves a list of all countries",
        "operationId": "getCountries",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/charities": {
      "get": {
        "tags": [
          "Charities"
        ],
        "summary": "Retrieves a list of all charities",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Charity"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/profile": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Adds a new user in Firebase Firestore",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "uid",
                "email"
              ],
              "properties": {
                "uid": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "UID or Email is missing"
          }
        }
      }
    },
    "/'user/profile/{uid}'": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Retrieves the details of a user profile",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "description": "unique user id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Updates a user profile",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "first",
                "last",
                "country"
              ],
              "properties": {
                "first": {
                  "type": "string"
                },
                "last": {
                  "type": "string"
                },
                "country": {
                  "type": "string"
                }
              }
            }
          },
          {
            "name": "uid",
            "in": "path",
            "description": "unique user id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/user/subscriptions": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "adds a user subscription",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "subscription",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "uid",
                "playerId",
                "name",
                "team",
                "teamName",
                "charityName",
                "charityId"
              ],
              "properties": {
                "uid": {
                  "type": "string"
                },
                "playerId": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "team": {
                  "type": "string"
                },
                "charityName": {
                  "type": "string"
                },
                "charityId": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "not found"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "updates a user subscription",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "subscription",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "uid",
                "playerId",
                "charityId",
                "charityName"
              ],
              "properties": {
                "uid": {
                  "type": "string"
                },
                "playerId": {
                  "type": "string"
                },
                "charityid": {
                  "type": "string"
                },
                "charityName": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "not found"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "removes a user subscription",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "uid",
                "playerId"
              ],
              "properties": {
                "uid": {
                  "type": "string"
                },
                "playerId": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/user/subscriptions/{uid}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Gets the subscriptions of the specified user",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "description": "unique user id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Subscription"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/user/subscriptions/{uid}/{playerId}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "returns whether or not a user is subscribed to a specific player",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "description": "unique user id",
            "required": true,
            "type": "string"
          },
          {
            "name": "playerId",
            "in": "path",
            "description": "unique player id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "boolean denoting whether or not the operation was successful",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "/user/matches/{uid}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Retrieves all matches involving a user's subscription players for the next month",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "description": "unique user id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation (empty array if user not found or no matches found)",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Match"
              }
            }
          }
        }
      }
    },
    "/user/participants/{uid}/{local}/{visitor}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Retrieves all players a user is subcribed to participating in a specific match",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "description": "unique user id",
            "required": true,
            "type": "string"
          },
          {
            "name": "local",
            "in": "path",
            "description": "home team id",
            "required": true,
            "type": "string"
          },
          {
            "name": "visitor",
            "in": "path",
            "description": "visitor team id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Profile": {
      "type": "object",
      "properties": {
        "birthday": {
          "type": "string",
          "example": "11/26/2018"
        },
        "country": {
          "type": "string",
          "example": "Canada"
        },
        "email": {
          "type": "string",
          "example": "myExampleEmail@gmail.com"
        },
        "first": {
          "type": "string",
          "example": "Ahmed"
        },
        "last": {
          "type": "string",
          "example": "Al-Zahrani"
        },
        "stats": {
          "type": "object",
          "properties": {
            "topScorer": {
              "type": "string",
              "example": ""
            },
            "allGoals": {
              "type": "array",
              "items": {
                "type": "object"
              }
            },
            "charities": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "topCharity": {
              "type": "string",
              "example": ""
            },
            "scorers": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "goals": {
              "type": "integer",
              "example": 0
            }
          }
        }
      }
    },
    "Subscription": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 158575,
          "description": "the id of the player in the subscription"
        },
        "charity": {
          "type": "string",
          "example": "Sightsavers"
        },
        "charityId": {
          "type": "integer",
          "example": 4
        },
        "name": {
          "type": "string",
          "example": "H.Kane"
        },
        "team": {
          "type": "integer",
          "example": 9406
        },
        "teamName": {
          "type": "string",
          "example": "Tottenham"
        },
        "goals": {
          "type": "integer",
          "example": 0
        },
        "time": {
          "type": "string",
          "format": "dateTime",
          "example": "Fri Aug 24 2018 01:35:51 GMT-0400 (Eastern Daylight Time)"
        }
      }
    },
    "Charity": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Unicef"
        },
        "id": {
          "type": "string",
          "example": "0009"
        },
        "description": {
          "type": "string",
          "example": "Unicef works in 190 countries and territories to save the lives of children"
        },
        "website": {
          "type": "string",
          "example": "https://www.unicef.org"
        },
        "imagePath": {
          "type": "string",
          "example": "assets/images/charities/Unicef.png"
        },
        "payment_info": {
          "type": "string",
          "example": "Unicef_PaymentInfo"
        }
      }
    },
    "Player": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 158575
        },
        "name": {
          "type": "string",
          "example": "H. Kane"
        },
        "age": {
          "type": "integer",
          "example": 25
        },
        "position": {
          "type": "string",
          "example": "Attacker"
        },
        "team": {
          "type": "integer",
          "example": 9406
        },
        "teamName": {
          "type": "string",
          "example": "Tottenham"
        },
        "league": {
          "type": "string",
          "example": "Premier League"
        },
        "number": {
          "type": "integer",
          "example": 10
        },
        "injured": {
          "type": "boolean",
          "example": false
        },
        "appearences": {
          "type": "integer",
          "example": 2
        },
        "goals": {
          "type": "integer",
          "example": 1
        },
        "assists": {
          "type": "integer",
          "example": 1
        },
        "yellowcards": {
          "type": "integer",
          "example": 0
        },
        "redcards": {
          "type": "integer",
          "example": 0
        }
      }
    },
    "Match": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 2378369
        },
        "comp_id": {
          "type": "integer",
          "example": 1204
        },
        "formatted_date": {
          "type": "string",
          "example": "18.08.2018"
        },
        "season": {
          "type": "string",
          "example": "2018/2019"
        },
        "week": {
          "type": "integer",
          "example": 2
        },
        "venue": {
          "type": "string",
          "example": "Wembley Stadium (London)"
        },
        "venue_id": {
          "type": "integer",
          "example": 87614
        },
        "venue_city": {
          "type": "string",
          "example": "London"
        },
        "status": {
          "type": "string",
          "example": "FT"
        },
        "timer": {
          "type": "string",
          "example": ""
        },
        "time": {
          "type": "string",
          "example": 840
        },
        "localteam_id": {
          "type": "integer",
          "example": 9406
        },
        "localteam_name": {
          "type": "string",
          "example": "Tottenham Hotspur"
        },
        "localteam_score": {
          "type": "integer",
          "example": 3
        },
        "visitorteam_id": {
          "type": "integer",
          "example": 9175
        },
        "visitorteam_name": {
          "type": "string",
          "example": "Fulham"
        },
        "visitorteam_score": {
          "type": "integer",
          "example": 1
        },
        "ht_score": {
          "type": "string",
          "example": "[1-0]"
        },
        "ft_score": {
          "type": "string",
          "example": "[3-1]"
        },
        "et_score": {
          "type": "string",
          "example": null
        },
        "penalty_local": {
          "type": "string",
          "example": null
        },
        "penalty_visitor": {
          "type": "string",
          "example": null
        },
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Event"
          }
        }
      }
    },
    "Event": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 24531251
        },
        "type": {
          "type": "string",
          "example": "goal"
        },
        "minute": {
          "type": "string",
          "example": 43
        },
        "extra_min": {
          "type": "string",
          "example": ""
        },
        "team": {
          "type": "string",
          "example": "localteam"
        },
        "player": {
          "type": "string",
          "example": "Lucas"
        },
        "player_id": {
          "type": "integer",
          "example": 141970
        },
        "assist": {
          "type": "string",
          "example": ""
        },
        "assist_id": {
          "type": "integer",
          "example": null
        },
        "result": {
          "type": "string",
          "example": "[1-0]"
        }
      }
    }
  }
}